backend/api/requirements.txt
fastapi
uvicorn
mysql-connector-python
pymongo
neo4j
